+++ io-classes/mtl/Control/Monad/Class/MonadEventlog/Trans.hs
@@ -18,39 +18,46 @@ import Control.Monad.Class.MonadEventlog
 instance MonadEventlog m => MonadEventlog (ContT r m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance MonadEventlog m => MonadEventlog (ExceptT e m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance (Monoid w, MonadEventlog m) => MonadEventlog (Lazy.RWST r w s m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance (Monoid w, MonadEventlog m) => MonadEventlog (Strict.RWST r w s m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance MonadEventlog m => MonadEventlog (Lazy.StateT s m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance MonadEventlog m => MonadEventlog (Strict.StateT s m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance (Monoid w, MonadEventlog m) => MonadEventlog (Lazy.WriterT w m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
+  flushEventLog = lift   flushEventLog
 
 -- | @since 0.1.0.0
 instance (Monoid w, MonadEventlog m) => MonadEventlog (Strict.WriterT w m) where
   traceEventIO  = lift . traceEventIO
   traceMarkerIO = lift . traceMarkerIO
-
+  flushEventLog = lift   flushEventLog